+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixing [#48](https://github.com/MultinetInteractive/EduAdmin-WordPress/issues/48), to allow users to choose "username" field
 - Login-code checks the given login field instead of email.
 - It is now possible to add your own translation directly in the plugin. (Again)
+- Added extra filter to course list (ajax) to skip "next" event if there isn't a public price name set.
 
 ## [0.9.9.2.40]
 ### Added